# PERF: DatetimeIndex comparison with Timestamp mismatched resos